requirements.txt
asgiref==3.7.2
attrs==22.1.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.3
coverage==6.5.0
fastavro==0.24.0
Flask==2.2.2
Flask-SQLAlchemy==3.0.2
flask-swagger==0.2.14
greenlet==2.0.1
grpcio==1.51.1
grpcio-tools==1.51.1
idna==3.6
iniconfig==1.1.1
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
packaging==21.3
pluggy==1.0.0
protobuf==4.25.3
psycopg2-binary==2.9.9
pulsar-client==2.10.2
py==1.11.0
PyDispatcher==2.0.6
pyparsing==3.0.9
pytest==7.1.3
pytz==2024.1
PyYAML==6.0
requests==2.28.2
six==1.16.0
SQLAlchemy==1.4.44
tomli==2.0.1
typing_extensions==4.10.0
urllib3==1.26.18
Werkzeug==2.2.2
aiopulsar-py==0.1.0a0
faker